Fix compiler warning on 64bit host.
git-svn-id: https://devel.aqbanking.de/svn/gwenhywfar/trunk@1510 70169cfe-8b10-0410-8925-dcb4b91034d8

+++ src/parser/text.c
@@ -275,7 +275,7 @@ int GWEN_Text_GetWordToBuffer(const char *src,
   } /* while */
 
   if (insideQuotes) {
-    DBG_ERROR(GWEN_LOGDOMAIN, "Missing \" after word (at %d: [%s])", src-savedSrc, savedSrc);
+    DBG_ERROR(GWEN_LOGDOMAIN, "Missing \" after word (at %d: [%s])", (int)(src-savedSrc), savedSrc);
     return -1;
   }
   /* check whether the source string was correctly terminated */
